requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
python-jose[cryptography]==3.3.0
redis[asyncio]==5.0.8
sentry-sdk[fastapi]==2.14.0
slowapi==0.1.9
limits==3.13.0
